requirements.txt
mysql-connector-python==8.0.28
mollie-api-python==3.1.1
pyyaml==6.0
